refactor(editor): House PopOutWindowKey in composables injectionKeys.ts (no-changelog)
Per Alex's review note on #34229: move the PopOutWindowKey definition from its
own per-key file into a conventional @n8n/composables/injectionKeys.ts module
(mirrors editor-ui's injectionKeys.ts), repoint useClipboard's import and
editor-ui's re-export shim at the @n8n/composables/injectionKeys subpath. Still
a single source; no barrel.
